Split Map into MapFromLbe and SanitizeAddress steps
